Revisionnnnn (#108)
* actualizacion color ticket

* detalle carrito

* anular compra admin

+++ src/main/java/pe/edu/pucp/fasticket/services/fidelizacion/FidelizacionService.java
@@ -441,7 +441,7 @@ public class FidelizacionService {
      */
     @Transactional
     public void revertirPuntosPorAnulacion(OrdenCompra orden) {
-        int puntosPorSol = Integer.parseInt(getConfig("PUNTOS_POR_MONEDA", "1"));
+        double puntosPorSol = Double.parseDouble(getConfig("PUNTOS_POR_MONEDA", "1"));
         int puntosARestar = (int) (orden.getTotal() * puntosPorSol);
 
         if (puntosARestar <= 0) return;
